requirements.txt
# Universal Testbook Generator - Demo Requirements

# Core Dependencies
streamlit>=1.32.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Document Processing
PyPDF2>=3.0.1
pdfplumber>=0.10.0

# AI/ML Libraries
openai>=1.30.0
pinecone-client>=3.0.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Utilities
loguru>=0.7.0
jinja2>=3.1.0

# Optional for enhanced features
streamlit-chat>=0.1.1